requirements.txt
alembic==0.8.10
appdirs==1.4.0
bcrypt==3.1.2
blinker==1.4
cffi==1.9.1
click==6.7
coverage==4.3.4
Flask==0.12
Flask-Bcrypt==0.7.1
Flask-DebugToolbar==0.10.1
Flask-Login==0.4.0
Flask-Migrate==2.0.3
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Testing==0.6.1
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.9.5
Mako==1.0.6
MarkupSafe==0.23
numexpr==2.6.2
numpy==1.12.0
packaging==16.8
pandas==0.19.2
pycparser==2.17
pyflakes==1.5.0
pyparsing==2.1.10
python-dateutil==2.6.0
python-decouple==3.0
python-editor==1.0.3
pytz==2016.10
six==1.10.0
SQLAlchemy==1.1.5
SQLAlchemy-Utils==0.32.12
tables==3.3.0
Werkzeug==0.11.15
